backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-genai
groq
python-multipart
PyPDF2